refactor: simplify build and packaging of card component
