+++ packages/components/card/bs-card-body.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsCardBodyCss } from './bs-card-body.css.js';
+import { BsCardBodyCss } from './css/bs-card-body.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsCardBody extends LitElement {
