+++ packages/components/card/bs-card-img.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html, css } from 'lit-element';
-import { BsCardImgCss } from './bs-card-img.css.js';
+import { BsCardImgCss } from './css/bs-card-img.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsCardImg extends LitElement {
