+++ packages/components/card/bs-card-links.js
@@ -1,7 +1,7 @@
 
 import { LitElement, html } from 'lit-element';
 import { configureCardLinks } from './helper/bs-card-links-helper.js';
-import { BsCardLinksCss } from './bs-card-links.css.js';
+import { BsCardLinksCss } from './css/bs-card-links.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsCardLinks extends LitElement {
