+++ packages/components/card/index.js
@@ -1,31 +1,17 @@
-
+export * from './css'
 export { BsCard } from './bs-card.js';
-export { BsCardCss } from './bs-card.css.js';
 export { BsCardText } from './bs-card-text.js';
-export { BsCardTextCss } from './bs-card-text.css.js';
 export { BsCardBody } from './bs-card-body.js';
-export { BsCardBodyCss } from './bs-card-body.css.js';
 export { BsCardTitle } from './bs-card-title.js';
-export { BsCardTitleCss } from './bs-card-title.css.js';
 export { BsCardOverlay } from './bs-card-overlay.js';
-export { BsCardOverlayCss } from './bs-card-overlay.css.js';
 export { BsCardSubtitle } from './bs-card-subtitle.js';
-export { BsCardSubtitleCss } from './bs-card-subtitle.css.js';
 export { BsCardLinks } from './bs-card-links.js';
-export { BsCardLinksCss } from './bs-card-links.css.js';
 export { BsCardLink } from './bs-card-link.js';
-export { BsCardLinkCss } from './bs-card-link.css.js';
 export { BsCardImg } from './bs-card-img.js';
-export { BsCardImgCss } from './bs-card-img.css.js';
 export { BsCardHeader } from './bs-card-header.js';
-export { BsCardHeaderCss } from './bs-card-header.css.js';
 export { BsCardGroup } from './bs-card-group.js';
-export { BsCardGroupCss } from './bs-card-group.css.js';
 export { BsCardFooter } from './bs-card-footer.js';
-export { BsCardFooterCss } from './bs-card-footer.css.js';
 export { BsCardDeck } from './bs-card-deck.js';
-export { BsCardDeckCss } from './bs-card-deck.css.js';
 export { BsCardColumns } from './bs-card-columns.js';
-export { BsCardColumnsCss } from './bs-card-columns.css.js';
 export { configureCardLinks } from './helper/bs-card-links-helper.js';
 export { configureFlushListGroupBorders } from './helper/bs-card-list-group-helper.js';
